requirements.txt
tomlkit>=0.14.0
python-dotenv>=1.2.1
